requirements.txt
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Fake data generation
faker==22.0.0

# Utils
python-dotenv==1.0.0
pandas==2.1.0

# Phase 3 — LLM Engine
langchain==0.1.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
huggingface-hub==0.20.0

neo4j==5.14.0